Fix ExtracurricularActivityService
