+++ app/src/main/java/edu/kmaooad/capstone23/activities/handlers/BulkCreateCoursesHandler.java
@@ -2,7 +2,6 @@ package edu.kmaooad.capstone23.activities.handlers;
 
 import edu.kmaooad.capstone23.activities.commands.BulkCreateCourses;
 import edu.kmaooad.capstone23.activities.dal.Course;
-import edu.kmaooad.capstone23.activities.dal.CourseRepository;
 import edu.kmaooad.capstone23.activities.events.BulkCoursesCreated;
 import edu.kmaooad.capstone23.activities.events.CourseCreated;
 import edu.kmaooad.capstone23.activities.services.CourseService;
