+++ app/src/main/java/edu/kmaooad/capstone23/activities/handlers/BulkDeleteCoursesHandler.java
@@ -2,7 +2,6 @@ package edu.kmaooad.capstone23.activities.handlers;
 
 import edu.kmaooad.capstone23.activities.commands.BulkDeleteCourses;
 import edu.kmaooad.capstone23.activities.dal.Course;
-import edu.kmaooad.capstone23.activities.dal.CourseRepository;
 import edu.kmaooad.capstone23.activities.events.BulkCoursesDeleted;
 import edu.kmaooad.capstone23.activities.events.CourseDeleted;
 import edu.kmaooad.capstone23.activities.services.CourseService;
