+++ app/src/main/java/edu/kmaooad/capstone23/activities/handlers/CreateRequestToJoinExtraActHandler.java
@@ -12,7 +12,6 @@ import edu.kmaooad.capstone23.activities.dal.RequestsRepository;
 import edu.kmaooad.capstone23.departments.events.RequestCreated;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
-import org.bson.types.ObjectId;
 
 import java.util.List;
