+++ app/src/main/java/edu/kmaooad/capstone23/activities/handlers/DeleteCourseHandler.java
@@ -1,7 +1,6 @@
 package edu.kmaooad.capstone23.activities.handlers;
 
 import edu.kmaooad.capstone23.activities.commands.DeleteCourse;
-import edu.kmaooad.capstone23.activities.dal.CourseRepository;
 import edu.kmaooad.capstone23.activities.events.CourseDeleted;
 import edu.kmaooad.capstone23.activities.services.CourseService;
 import edu.kmaooad.capstone23.common.CommandHandler;
